feat: add base modularization
