+++ ichigo/asr/arch/__init__.py
@@ -1,0 +1,5 @@
+from .s2r import Speech2Rep
+from .quantizer import Quantizer
+from .r2t import Rep2Text
+
+__all__ = ["Speech2Rep", "Quantizer", "Rep2Text"]
