--- ichigo/asr/models/__init__.py
@@ -1,4 +1,0 @@
-from .factory import make_vq_model
-from .vq_transformer import RQBottleneckTransformer
-
-__all__ = ["make_vq_model", "RQBottleneckTransformer"]
